backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
djoser>=2.2,<3.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.2,<25.0